Refactor test structure to move massive fuzz tests into their own module
